Use uintmax_t for AmmoniteId

+++ src/include/ammonite/types.hpp
@@ -2,6 +2,7 @@
 #define TYPES
 
 #include <climits>
+#include <cstdint>
 #include <semaphore>
 #include <vector>
 
@@ -9,6 +10,6 @@ using AmmoniteKeyCallback = void (*)(const std::vector<int>& keycodes, int actio
 using AmmoniteWork = void (*)(void* userPtr);
 
 using AmmoniteGroup = std::counting_semaphore<INT_MAX>;
-using AmmoniteId = unsigned int;
+using AmmoniteId = uintmax_t;
 
 #endif
